PRICE FIX #110 #107

+++ RogueLikeMADY/src/main/java/com/mady/utils/entities/factories/items/Shop.java
@@ -151,7 +151,7 @@ public class Shop extends Map {
     private void addItem(PaidChest chest,Player player) {
         AbstractStuffItem item = chest.selectItem(player);
         Price price = new Price(item);
-        chest.setPrice(price.getPrice()+(int) (price.getPrice()));
+        chest.setPrice(price.getPrice()+ price.getPrice());
         //it.add(item);
         items.put(item, price);
     }
